+++ src/javascript/config.js
@@ -11,9 +11,88 @@ Paris.config = {
     // You can add any Algolia index here, to be able to use it in the quick-access module and search-results template
     "indexes": {
       "global": "recette_ParisFront",
-      "persons": "recette_Elus"
+      "persons": "recette_Elus",
+      "equipments": "Service_recette",
     },
-
+    "main_facets": ["Bibliothèques", "Mairies d'arrondissement", "Principaux parcs, jardins et squares", "Piscines", "Tennis"],
+    "main_facets_icon": [
+      {
+        "facet": "Piscines",
+        "icon": "swimming-pool"
+      },
+      {
+        "facet": "Bibliothèques",
+        "icon": "library"
+      },
+      {
+        "facet": "Médiathèques et ludothèques",
+        "icon": "library"
+      },
+      {
+        "facet": "Principaux parcs, jardins et squares",
+        "icon": "park"
+      },
+      {
+        "facet": "Jardins, mails, promenades",
+        "icon": "park"
+      },
+      {
+        "facet": "Squares",
+        "icon": "park"
+      },
+      {
+        "facet": "Marchés alimentaires et spécialisés",
+        "icon": "market"
+      },
+      {
+        "facet": "Marchés spécialisés",
+        "icon": "market"
+      },
+      {
+        "facet": "Ateliers Beaux-Arts",
+        "icon": "art-workshop"
+      },
+      {
+        "facet": "Centres Paris Anim",
+        "icon": "animation"
+      },
+      {
+        "facet": "Conservatoires municipaux",
+        "icon": "conservatoire"
+      },
+      {
+        "facet": "Mairies d'arrondissement",
+        "icon": "city-hall"
+      },
+      {
+        "facet": "Hôtel de Ville",
+        "icon": "city-hall"
+      },
+      {
+        "facet": "Maisons des associations",
+        "icon": "associations"
+      },
+      {
+        "facet": "Point Paris Emploi",
+        "icon": "jobs"
+      },
+      {
+        "facet": "Musées municipaux",
+        "icon": "museum"
+      },
+      {
+        "facet": "Tennis",
+        "icon": "tennis"
+      },
+      {
+        "facet": "Théàtres de la Ville et établissements soutenus",
+        "icon": "culture"
+      },
+      {
+        "facet": "Théàtres pour enfants",
+        "icon": "culture"
+      }
+    ],
     "url": {
       "api_popular_searches": "http://r7.paris-fr-api.lestudio.mx/AlgoliaStats/getMostSearchedKeywords"
     }
@@ -42,10 +121,18 @@ Paris.config = {
   "captcha": {
     "key": "6Lf2DQoTAAAAAKmk3wEFCZuK3FqG00SlQM3o6Yvp"
   },
-  "mapbox": {
+  "leaflet": {
     "accessToken": "pk.eyJ1IjoicGFyaXNudW1lcmlxdWUiLCJhIjoiY2loZG1vMnYyMDAzNnY0a3FvNG1nNG55biJ9.MP1qcHkEecFGqSTs9gg7cw",
     "defaultCenter": {"lat": 48.856578, "lon": 2.351828},
-    "defaultBounds": [[48.902157, 2.46976], [48.815575, 2.224122]],
-    "styleLayer": "mapbox://styles/parisnumerique/cis1rkqoj000khinpthppoaqd"
+    "minZoom": 11,
+    "maxBounds": [[49.2,2.56], [48.76,2.11]],
+    "tileLayer": "http://filer.paris.fr/leaflet/paris2/{z}/{x}/{y}.png",
+    "tileLayerAttribution": "&copy; <a href=\"http://osm.org/copyright\">OpenStreetMap</a> contributors"
+  },
+  "search": {
+    "paris_coordinates" : {
+      "lat": 48.856579,
+      "lng": 2.349272
+    }
   }
 };